src/api/authentification/requirements.txt
fastapi==0.116.1
uvicorn==0.30.6
pydantic==2.11.7
pyjwt==2.10.1
python-multipart==0.0.20